+++ test.py
@@ -50,11 +50,11 @@ def read_input(in_file):
 def test(fname, eps = 1e-5):
   args, X, F2d, F1d, Y = read_input(fname)
   conv = nn.Conv2d(in_channels=int(args[3]), out_channels=int(args[3]), kernel_size=int(args[4]), stride=int(args[10]), groups=int(args[3]), bias=False)
-  conv.weight = torch.nn.Parameter(F2d)
+  conv.weight = torch.nn.Parameter(F2d.float())
   point_conv = nn.Conv2d(in_channels=int(args[3]), out_channels=int(args[9]), kernel_size=1, bias=False)
-  point_conv.weight = torch.nn.Parameter(F1d)
+  point_conv.weight = torch.nn.Parameter(F1d.float())
   depthwise_separable_conv = torch.nn.Sequential(conv, point_conv)
-  out = depthwise_separable_conv(X.double())
+  out = depthwise_separable_conv(X.float())
   assert (out - Y).abs().max().item() <= eps, "FAILED CORRECTNESS"
   print("Passed correctness!")
